requirements.txt
Flask==2.0.2
grpcio
requests==2.26.0
opentelemetry-api==1.5.0
opentelemetry-exporter-jaeger==1.5.0
opentelemetry-sdk==1.5.0
